requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
asttokens==3.0.0
asyncio==4.0.0
bidict==0.23.1
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cmake==4.1.2
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
dotenv==0.9.9
executing==2.2.1
fastapi==0.121.3
fastembed-gpu==0.7.3
filelock==3.20.0
flatbuffers==25.9.23
fs==2.4.16
fsspec==2025.10.0
google-auth==2.42.1
google-genai==1.47.0
groq==0.33.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
ipykernel==7.1.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.9.1
loguru==0.7.3
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mmh3==5.2.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==1.26.4
onnx==1.17.0
onnxruntime-gpu==1.23.2
packaging==25.0
parso==0.8.5
pathlib==1.0.1
pillow==11.3.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
protobuf==6.33.0
psutil==7.1.2
pure_eval==0.2.3
py_rust_stemmers==0.1.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pymongo==4.15.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.12.3
python-socketio==5.14.3
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.10.23
requests==2.32.5
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uv==0.9.7
uvicorn==0.38.0
wcwidth==0.2.14
websockets==15.0.1
wheel==0.45.1
win32_setctime==1.2.0
wsproto==1.3.2
